Fix alpha and color space issues in render stamp and color mapping code

+++ src/appleseed/foundation/resources/logo/appleseed-seeds-16.h
@@ -29,9 +29,6 @@
 #ifndef APPLESEED_FOUNDATION_RESOURCES_LOGO_APPLESEED_SEEDS_16_H
 #define APPLESEED_FOUNDATION_RESOURCES_LOGO_APPLESEED_SEEDS_16_H
 
-// appleseed.foundation headers.
-#include "foundation/platform/types.h"
-
 // appleseed.main headers.
 #include "main/dllsymbol.h"
 
@@ -42,13 +39,18 @@ namespace foundation
 {
 
 //
-// appleseed seeds logo, 16x15, 8-bit per channel, RGBA.
+// appleseed "seeds" logo:
+//
+//   16x15 pixels
+//   32-bit floating point RGBA
+//   linear RGB
+//   premultiplied alpha
 //
 
 const size_t appleseed_seeds_16_width = 16;
 const size_t appleseed_seeds_16_height = 15;
 
-APPLESEED_DLLSYMBOL extern const uint8 appleseed_seeds_16[];
+APPLESEED_DLLSYMBOL extern const float appleseed_seeds_16[];
 
 }       // namespace foundation
